🔥 Supprime l'impression du formulaire DNRTI
Et simplification du code
