+++ mon-entreprise/package.json
@@ -49,7 +49,6 @@
 		"react-router-hash-link": "^1.2.2",
 		"react-spring": "=8.0.27",
 		"react-syntax-highlighter": "^10.1.1",
-		"react-to-print": "^2.5.1",
 		"react-transition-group": "^2.2.1",
 		"recharts": "^1.8.5",
 		"reduce-reducers": "^1.0.4",
