+++ mon-entreprise/source/sites/mon-entreprise.fr/pages/Gérer/Home.tsx
@@ -317,7 +317,7 @@ export const CompanySection = ({ company }: CompanySectionProps) => {
 					<ScrollToTop />
 					<Overlay>
 						<Trans i18nKey="gérer.entreprise.dirigeant">
-							<h2> Êtes-vous dirigeant majoritaire ? </h2>
+							<h2>Êtes-vous dirigeant majoritaire ?</h2>
 							<p>
 								Si vous êtes administrateur majoritaire ou si vous faites partie
 								d'un conseil d'administration majoritaire, vous n'aurez pas le
